Feature/roles

+++ app/Http/Controllers/PostControllerApi.php
@@ -5,13 +5,13 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
 use App\Post;
+use Gate;
 
 class PostControllerApi extends Controller
 {
 
     public function index(): JsonResponse
     {
-
         $model = Post::all();
         if (!count($model)) {
             return $this->jsonResponse('Not found posts', 400);
@@ -21,9 +21,13 @@ class PostControllerApi extends Controller
 
     public function store(Request $request): JsonResponse
     {
-
         $post = $this->trim($request->post());
         $model = new Post;
+
+        if (Gate::denies('create', $model)) {
+            return $this->jsonResponse(['You can not create posts.'], 400);
+        }
+
         $model->addPost($post);
         $model->addValidate([
             'user_id' => ['required',],
@@ -33,16 +37,15 @@ class PostControllerApi extends Controller
         $result = $model->validate();
         if (count($result)) {
             return $this->jsonResponse($result, 400);
-        } else {
-            $model->timestamps = false;
-            $model->save();
         }
+        $model->timestamps = false;
+        $model->save();
+
         return $this->jsonResponse(['Post successfully added', $model], 201);
     }
 
     public function show($id): JsonResponse
     {
-
         $data = Post::find($id);
         if (!empty($data)) {
             return $this->jsonResponse($data, 200);
@@ -52,35 +55,41 @@ class PostControllerApi extends Controller
 
     public function update(Request $request, $id): JsonResponse
     {
-
         $model = new Post;
         $model = $model->find($id);
+
+        if (Gate::denies('update', $model)) {
+            return $this->jsonResponse(['You can not update posts.'], 400);
+        }
+
         if (empty($model)) {
             return $this->jsonResponse(['Alert' => ['Not found post']], 400);
-        } else {
-            $post = $this->trim($request->post());
-            $model->addPost($post);
-            $result = $model->validate();
-            if (count($result)) {
-                return $this->jsonResponse($result, 400);
-            } else {
-                $model->timestamps = false;
-                $model->save();
-            }
         }
+
+        $post = $this->trim($request->post());
+        $model->addPost($post);
+        $result = $model->validate();
+        if (count($result)) {
+            return $this->jsonResponse($result, 400);
+        }
+        $model->timestamps = false;
+        $model->save();
+
         return $this->jsonResponse(['Post successfully updated', $model], 201);
 
     }
 
     public function destroy(int $id): JsonResponse
     {
-
         $model = Post::find($id);
         if (empty($model)) {
             return $this->jsonResponse(['Alert' => ['Not found post']], 400);
-        } else {
-            $model = Post::destroy($id);
         }
+        if (Gate::denies('update', $model)) {
+            return $this->jsonResponse(['You can not delete posts.'], 400);
+        }
+        $model = Post::destroy($id);
+
         if ($model === 0) {
             return $this->jsonResponse(['Alert' => ['Unknown error']], 400);
         }
